+++ src/L10NSharp.Tests/LocalizationManagerTestsBase.cs
@@ -216,19 +216,19 @@ namespace L10NSharp.Tests
 
 				Assert.AreEqual("My Own English String",
 					ProxyLocalizationManager.MyOwnGetString("myOwn.English.String.Id", "My Own English String"));
-				
+
 				Assert.AreEqual("My Own English String (with comment)",
 					ProxyLocalizationManager.MyOwnGetString("myOwn.English.String.Id.With.Comment",
 					"My Own English String (with comment)",
 					"This is used to test the case where MyOwnGetString is passed as an extra method to use for extraction."));
-				
+
 				Assert.AreEqual("Click me",
 					ProxyLocalizationManager.MyOwnGetString("myDlg.btnClickMe.Text", "Click me",
 					"This is the text from the third version of MyOwnGetString.",
 					"Click this thingy to do stuff.", "Ctrl-T", btnClickMe));
-				
+
 				Assert.AreEqual("String to Localize", "String to Localize".Localize());
-				
+
 				Assert.AreEqual("Another String to Localize", "Another String to Localize".Localize("With.Id.And.Comment",
 					"This is used to test the case where Localize is passed as an extra method to use for extraction."));
 			}
@@ -354,6 +354,32 @@ namespace L10NSharp.Tests
 
 				Assert.AreEqual(expectedResult, result);
 				Assert.AreEqual(expectedLanguage, languageFound);
+			}
+		}
+
+		[Test]
+		public void GetString_WithOneShotEnumerable_UsesPreferredLanguages()
+		{
+			using (var folder = new TempFolder())
+			{
+				SetupManager(folder);
+
+				// Wrap a single IEnumerator so the sequence can only be consumed once.
+				IEnumerable<string> WrapSingleUse(IEnumerator<string> e)
+				{
+					while (e.MoveNext())
+						yield return e.Current;
+				}
+
+				var oneShotLangIds = WrapSingleUse(
+					new List<string> { "fr", "en" }.GetEnumerator());
+
+				// SUT
+				var result = LocalizationManager.GetString("blahId", "blahInEnglishCode",
+					"comment", oneShotLangIds, out var languageFound);
+
+				Assert.AreEqual("blahInFrench", result);
+				Assert.AreEqual("fr", languageFound);
 			}
 		}